Add counting box that can self update
Add a counting box that's intended to self update. Strangely this
current commit is subject to a bug. The elements update themselves
correctly when clicked. But for some reason, the click event does
not give the correct updated values when fired from javascript.
Strangely clicking from the browser does correctly update the
values.